requirements.txt
python-telegram-bot==20.3
telethon
python-dotenv
fastapi
uvicorn
